LayoutTests/svg/animations/animate-dynamic-update-attributeName-expected.txt

SVG 1.1 dynamic animation tests

Test behavior on dynamic-update of attributeName

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS color.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS color.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS rect.style.color is ""
PASS color.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 128
PASS color.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS rect.style.color is ""
PASS color.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 128
PASS color.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS rect.style.color is ""
PASS rect.style.fill is ""
PASS successfullyParsed is true

TEST COMPLETE
